test-suite/requirements.txt
PyYAML
requests
boto3
rich
pydantic
numpy
pandas
flask
matplotlib
sqlalchemy
